Refactored testing to fix pylint errors
